src/requirements.txt
# AWS SDK
boto3==1.34.51

# JWT validation
python-jose[cryptography]==3.3.0

# HTTP requests
requests==2.31.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
hypothesis==6.98.3
moto==5.0.1

# Utilities
python-dateutil==2.8.2
